fix(server): swagger authorize parameters in case, when authorize attribute is attached to controller
