+++ GomokuServer/src/GomokuServer.Api/GomokuServer.Api.csproj
@@ -10,6 +10,13 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Attributes\**" />
+    <Content Remove="Attributes\**" />
+    <EmbeddedResource Remove="Attributes\**" />
+    <None Remove="Attributes\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Asp.Versioning.Mvc" Version="8.1.0" />
     <PackageReference Include="Asp.Versioning.Mvc.ApiExplorer" Version="8.1.0" />
     <PackageReference Include="DotNetEnv" Version="3.1.1" />
@@ -25,10 +32,6 @@
   <ItemGroup>
     <ProjectReference Include="..\GomokuServer.Application\GomokuServer.Application.csproj" />
     <ProjectReference Include="..\GomokuServer.Infrastructure\GomokuServer.Infrastructure.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Attributes\" />
   </ItemGroup>
 
 </Project>
